@@ -9,14 +9,15 @@
     "build": "webpack --mode production"
   },
   "dependencies": {
+    "@ant-design/icons": "^6.0.0",
     "ag-grid-react": "^33.3.2",
-    "antd": "^3.26.20",
-    "react": "^16.14.0",
-    "react-dom": "^16.14.0"
+    "antd": "^5.26.5",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1"
   },
   "devDependencies": {
-    "@types/react": "^16.14.0",
-    "@types/react-dom": "^16.9.0",
+    "@types/react": "^18.3.23",
+    "@types/react-dom": "^18.3.7",
     "css-loader": "^3.6.0",
     "html-webpack-plugin": "^5.6.3",
     "less": "^4.3.0",
