@@ -1,14 +1,13 @@
-import 'antd/dist/antd.css';
-import React, { useState, DragEvent, FC, StrictMode, useEffect } from 'react';
-import ReactDOM from 'react-dom';
-import ResizableDraggablePanel from './components/ResizableDraggablePanel';
-import TermsIcon from './Icons/TermsIcon';
-import AboutIcon from './Icons/AboutIcon';
-import FruitViewIcon from './Icons/FruitViewIcon';
-import { MainWorkspace } from './components/MainWorkspace';
-import LoginComponent from './components/LoginComponent';
-import UserProfile from './components/UserProfile';
-import { panelList } from './panelList';
+import React, { useState, DragEvent, FC, StrictMode, useEffect } from "react";
+import ReactDOM from "react-dom/client";
+import ResizableDraggablePanel from "./components/ResizableDraggablePanel";
+import TermsIcon from "./Icons/TermsIcon";
+import AboutIcon from "./Icons/AboutIcon";
+import FruitViewIcon from "./Icons/FruitViewIcon";
+import { MainWorkspace } from "./components/MainWorkspace";
+import LoginComponent from "./components/LoginComponent";
+import UserProfile from "./components/UserProfile";
+import { panelList } from "./panelList";
 
 /**
  * Represents an open panel's state and position.
@@ -31,8 +30,8 @@ type OpenPanel = {
 const getDefaultPanelPosition = (count: number) => ({
   x: 60 + count * 40,
   y: 60 + count * 40,
-  width: 700,  
-  height: 420, 
+  width: 700,
+  height: 420,
 });
 
 const GRID_ROWS = 2;
@@ -69,11 +68,11 @@ const NAV_BAR_HEIGHT = 56; // px, must match your nav bar minHeight
  * Checks if the user is logged in.
  * @returns {boolean}
  */
-const isLoggedIn = () => localStorage.getItem('isLoggedIn') === 'true';
+const isLoggedIn = () => localStorage.getItem("isLoggedIn") === "true";
 
 const INACTIVITY_LIMIT = 5 * 60 * 1000; // 5 minutes
 
-const THEME_KEY = 'theme'; // localStorage key
+const THEME_KEY = "theme"; // localStorage key
 
 /**
  * Gets the initial theme from localStorage or prompts the user.
@@ -81,10 +80,12 @@ const THEME_KEY = 'theme'; // localStorage key
  */
 const getInitialTheme = () => {
   const stored = localStorage.getItem(THEME_KEY);
-  if (stored === 'dark' || stored === 'light') return stored;
+  if (stored === "dark" || stored === "light") return stored;
   // Ask user if not set
-  const userPref = window.confirm('Use dark theme? Click OK for dark, Cancel for light.');
-  const theme = userPref ? 'dark' : 'light';
+  const userPref = window.confirm(
+    "Use dark theme? Click OK for dark, Cancel for light."
+  );
+  const theme = userPref ? "dark" : "light";
   localStorage.setItem(THEME_KEY, theme);
   return theme;
 };
@@ -97,8 +98,10 @@ const App: FC = () => {
   const [dragNavPanelKey, setDragNavPanelKey] = useState<string | null>(null);
   const [navOpen, setNavOpen] = useState<boolean>(false);
   const [containerSize, setContainerSize] = useState({ width: 0, height: 0 });
-  const [dropCell, setDropCell] = useState<{ row: number; col: number } | null>(null);
-  const [theme, setTheme] = useState<'dark' | 'light'>(getInitialTheme());
+  const [dropCell, setDropCell] = useState<{ row: number; col: number } | null>(
+    null
+  );
+  const [theme, setTheme] = useState<"dark" | "light">(getInitialTheme());
 
   // Drag from nav: set key in dataTransfer
   /**
@@ -107,13 +110,16 @@ const App: FC = () => {
    */
   const onNavDragStart = (key: string) => (e: DragEvent<HTMLLIElement>) => {
     setDragNavPanelKey(key);
-    e.dataTransfer.setData('panelKey', key);
+    e.dataTransfer.setData("panelKey", key);
   };
 
   /**
    * Called by MainWorkspace to update container size and drop cell.
    */
-  const handleGridDropInfo = (info: { cell: { row: number; col: number } | null, size: { width: number; height: number } }) => {
+  const handleGridDropInfo = (info: {
+    cell: { row: number; col: number } | null;
+    size: { width: number; height: number };
+  }) => {
     setDropCell(info.cell);
     setContainerSize(info.size);
   };
@@ -123,12 +129,15 @@ const App: FC = () => {
    */
   const onMainDrop = (e: DragEvent<HTMLDivElement>) => {
     e.preventDefault();
-    const key = e.dataTransfer.getData('panelKey');
+    const key = e.dataTransfer.getData("panelKey");
     if (!key) return;
-    const panelDef = panelList.find(p => p.key === key);
+    const panelDef = panelList.find((p) => p.key === key);
     if (!panelDef) return;
     const id = `${key}-${Date.now()}`;
-    let x = 60, y = NAV_BAR_HEIGHT + 10, width = 700, height = 420;
+    let x = 60,
+      y = NAV_BAR_HEIGHT + 10,
+      width = 700,
+      height = 420;
     if (dropCell && containerSize.width && containerSize.height) {
       // Subtract nav bar height from available height for grid
       const availableHeight = containerSize.height - NAV_BAR_HEIGHT;
@@ -172,19 +181,15 @@ const App: FC = () => {
    * Closes a panel by id.
    */
   const handleClose = (id: string) => {
-    setOpenPanels(openPanels.filter(p => p.id !== id));
+    setOpenPanels(openPanels.filter((p) => p.id !== id));
   };
 
   /**
    * Moves a panel by delta x and y.
    */
   const handlePanelMove = (id: string, dx: number, dy: number) => {
-    setOpenPanels(panels =>
-      panels.map(p =>
-        p.id === id
-          ? { ...p, x: p.x + dx, y: p.y + dy }
-          : p
-      )
+    setOpenPanels((panels) =>
+      panels.map((p) => (p.id === id ? { ...p, x: p.x + dx, y: p.y + dy } : p))
     );
   };
 
@@ -192,10 +197,14 @@ const App: FC = () => {
    * Resizes a panel by delta width and height.
    */
   const handlePanelResize = (id: string, dw: number, dh: number) => {
-    setOpenPanels(panels =>
-      panels.map(p =>
+    setOpenPanels((panels) =>
+      panels.map((p) =>
         p.id === id
-          ? { ...p, width: Math.max(200, p.width + dw), height: Math.max(100, p.height + dh) }
+          ? {
+              ...p,
+              width: Math.max(200, p.width + dw),
+              height: Math.max(100, p.height + dh),
+            }
           : p
       )
     );
@@ -208,20 +217,20 @@ const App: FC = () => {
     const resetTimer = () => {
       if (timer) clearTimeout(timer);
       timer = setTimeout(() => {
-        localStorage.removeItem('isLoggedIn');
-        window.dispatchEvent(new Event('login-success'));
+        localStorage.removeItem("isLoggedIn");
+        window.dispatchEvent(new Event("login-success"));
       }, INACTIVITY_LIMIT);
     };
 
-    const activityEvents = ['mousemove', 'keydown', 'mousedown', 'touchstart'];
-    activityEvents.forEach(event =>
+    const activityEvents = ["mousemove", "keydown", "mousedown", "touchstart"];
+    activityEvents.forEach((event) =>
       window.addEventListener(event, resetTimer)
     );
     resetTimer();
 
     return () => {
       if (timer) clearTimeout(timer);
-      activityEvents.forEach(event =>
+      activityEvents.forEach((event) =>
         window.removeEventListener(event, resetTimer)
       );
     };
@@ -229,7 +238,7 @@ const App: FC = () => {
 
   // Set theme class on body or root
   useEffect(() => {
-    document.body.classList.remove('theme-dark', 'theme-light');
+    document.body.classList.remove("theme-dark", "theme-light");
     document.body.classList.add(`theme-${theme}`);
   }, [theme]);
 
@@ -237,58 +246,64 @@ const App: FC = () => {
    * Toggles the application theme between dark and light.
    */
   const handleThemeToggle = () => {
-    setTheme(prev => {
-      const next = prev === 'dark' ? 'light' : 'dark';
+    setTheme((prev) => {
+      const next = prev === "dark" ? "light" : "dark";
       localStorage.setItem(THEME_KEY, next);
       return next;
     });
   };
 
   return (
-    <div className={`app-root theme-${theme}`} style={{ display: 'flex', height: '100vh' }}>
+    <div
+      className={`app-root theme-${theme}`}
+      style={{ display: "flex", height: "100vh" }}
+    >
       {/* Navigation Bar */}
       {navOpen && (
         <nav
           style={{
             width: 90, // Increased width
-            background: '#232b3e',
-            padding: '0.5rem 0.25rem',
-            borderRight: '1px solid #3e4a6b',
-            display: 'flex',
-            flexDirection: 'column',
-            alignItems: 'center',
+            background: "#232b3e",
+            padding: "0.5rem 0.25rem",
+            borderRight: "1px solid #3e4a6b",
+            display: "flex",
+            flexDirection: "column",
+            alignItems: "center",
             minWidth: 90, // Increased minWidth
-            boxSizing: 'border-box',
+            boxSizing: "border-box",
           }}
         >
-          <ul style={{
-            listStyle: 'none',
-            padding: 0,
-            margin: 0,
-            width: '100%',
-            display: 'flex',
-            flexDirection: 'column',
-            alignItems: 'center',
-          }}>
+          <ul
+            style={{
+              listStyle: "none",
+              padding: 0,
+              margin: 0,
+              width: "100%",
+              display: "flex",
+              flexDirection: "column",
+              alignItems: "center",
+            }}
+          >
             {panelList.map((panel) => (
               <li
                 key={panel.key}
                 style={{
                   marginBottom: 16,
-                  cursor: 'grab',
-                  fontWeight: 'normal',
-                  background: dragNavPanelKey === panel.key ? '#353b4a' : undefined,
+                  cursor: "grab",
+                  fontWeight: "normal",
+                  background:
+                    dragNavPanelKey === panel.key ? "#353b4a" : undefined,
                   padding: 8,
                   borderRadius: 10,
-                  display: 'flex',
-                  flexDirection: 'column',
-                  alignItems: 'center',
-                  justifyContent: 'center',
+                  display: "flex",
+                  flexDirection: "column",
+                  alignItems: "center",
+                  justifyContent: "center",
                   fontSize: 13,
-                  color: '#e0e0e0',
-                  width: '100%',
-                  transition: 'background 0.2s',
-                  textAlign: 'center', // Center text
+                  color: "#e0e0e0",
+                  width: "100%",
+                  transition: "background 0.2s",
+                  textAlign: "center", // Center text
                   minHeight: 64,
                 }}
                 draggable
@@ -297,22 +312,24 @@ const App: FC = () => {
                 title={panel.title}
               >
                 <span style={{ marginBottom: 4 }}>
-                  {panel.key === 'fruitbook' ? (
+                  {panel.key === "fruitbook" ? (
                     <TermsIcon />
-                  ) : panel.key === 'fruitview' ? (
+                  ) : panel.key === "fruitview" ? (
                     <FruitViewIcon />
-                  ) : panel.key === 'about' ? (
+                  ) : panel.key === "about" ? (
                     <AboutIcon />
                   ) : null}
                 </span>
-                <span style={{
-                  width: '100%',
-                  textAlign: 'center', // Center text
-                  fontSize: 13,
-                  fontWeight: 500,
-                  lineHeight: 1.2,
-                  wordBreak: 'break-word',
-                }}>
+                <span
+                  style={{
+                    width: "100%",
+                    textAlign: "center", // Center text
+                    fontSize: 13,
+                    fontWeight: 500,
+                    lineHeight: 1.2,
+                    wordBreak: "break-word",
+                  }}
+                >
                   {panel.title}
                 </span>
               </li>
@@ -332,60 +349,76 @@ const App: FC = () => {
         <main
           style={{
             flex: 1,
-            position: 'relative',
-            background: '#232b3e',
-            overflow: 'hidden',
-            height: '100%',
-            width: '100%',
+            position: "relative",
+            background: "#232b3e",
+            overflow: "hidden",
+            height: "100%",
+            width: "100%",
           }}
         >
           {/* Top nav branding */}
           <div
             style={{
-              width: '100%',
-              background: 'linear-gradient(90deg, #2b3556 0%, #3e4a6b 100%)',
-              color: '#fff',
-              padding: '0.5rem 1.5rem',
+              width: "100%",
+              background: "linear-gradient(90deg, #2b3556 0%, #3e4a6b 100%)",
+              color: "#fff",
+              padding: "0.5rem 1.5rem",
               fontWeight: 600,
               fontSize: 20,
               letterSpacing: 1,
-              position: 'sticky',
+              position: "sticky",
               top: 0,
               zIndex: 2000,
-              display: 'flex',
-              alignItems: 'center',
-              boxShadow: '0 2px 8px #0002',
+              display: "flex",
+              alignItems: "center",
+              boxShadow: "0 2px 8px #0002",
               minHeight: NAV_BAR_HEIGHT,
-              borderBottom: '1px solid #3e4a6b',
+              borderBottom: "1px solid #3e4a6b",
             }}
           >
             {/* Hamburger/X icon */}
             <button
               onClick={() => setNavOpen((v) => !v)}
               style={{
-                background: 'transparent',
-                border: 'none',
-                color: '#fff',
+                background: "transparent",
+                border: "none",
+                color: "#fff",
                 fontSize: 26,
-                cursor: 'pointer',
+                cursor: "pointer",
                 marginRight: 20,
-                display: 'flex',
-                alignItems: 'center',
+                display: "flex",
+                alignItems: "center",
                 padding: 0,
                 height: 40,
                 width: 40,
                 borderRadius: 8,
-                transition: 'background 0.2s',
-                boxShadow: navOpen ? '0 2px 8px #0002' : undefined,
+                transition: "background 0.2s",
+                boxShadow: navOpen ? "0 2px 8px #0002" : undefined,
               }}
               aria-label="Toggle navigation"
             >
-              <span style={{ display: 'inline-block', width: 28, height: 28 }}>
+              <span style={{ display: "inline-block", width: 28, height: 28 }}>
                 {navOpen ? (
                   // X icon
                   <svg width="28" height="28" viewBox="0 0 28 28">
-                    <line x1="7" y1="7" x2="21" y2="21" stroke="#fff" strokeWidth="2.5" strokeLinecap="round"/>
-                    <line x1="21" y1="7" x2="7" y2="21" stroke="#fff" strokeWidth="2.5" strokeLinecap="round"/>
+                    <line
+                      x1="7"
+                      y1="7"
+                      x2="21"
+                      y2="21"
+                      stroke="#fff"
+                      strokeWidth="2.5"
+                      strokeLinecap="round"
+                    />
+                    <line
+                      x1="21"
+                      y1="7"
+                      x2="7"
+                      y2="21"
+                      stroke="#fff"
+                      strokeWidth="2.5"
+                      strokeLinecap="round"
+                    />
                   </svg>
                 ) : (
                   // Hamburger icon
@@ -398,16 +431,18 @@ const App: FC = () => {
               </span>
             </button>
             {/* App title */}
-            <span style={{
-              fontFamily: 'monospace',
-              fontWeight: 700,
-              fontSize: 22,
-              letterSpacing: 2,
-              color: '#fff',
-              textShadow: '0 1px 2px #0006',
-              userSelect: 'none',
-              textTransform: 'uppercase',
-            }}>
+            <span
+              style={{
+                fontFamily: "monospace",
+                fontWeight: 700,
+                fontSize: 22,
+                letterSpacing: 2,
+                color: "#fff",
+                textShadow: "0 1px 2px #0006",
+                userSelect: "none",
+                textTransform: "uppercase",
+              }}
+            >
               fruteria
             </span>
             {/* Spacer to push UserProfile to the right */}
@@ -416,8 +451,8 @@ const App: FC = () => {
             <div style={{ marginRight: 32 }}>
               <UserProfile
                 onLogout={() => {
-                  localStorage.removeItem('isLoggedIn');
-                  window.dispatchEvent(new Event('login-success'));
+                  localStorage.removeItem("isLoggedIn");
+                  window.dispatchEvent(new Event("login-success"));
                 }}
                 onThemeToggle={handleThemeToggle}
                 theme={theme}
@@ -425,8 +460,11 @@ const App: FC = () => {
             </div>
           </div>
           {openPanels.length === 0 ? (
-            <div style={{ color: '#888', textAlign: 'center', marginTop: '2rem' }}>
-              No panels open.<br />
+            <div
+              style={{ color: "#888", textAlign: "center", marginTop: "2rem" }}
+            >
+              No panels open.
+              <br />
               Drag one from the navigation bar.
             </div>
           ) : (
@@ -456,8 +494,8 @@ const Root: React.FC = () => {
   useEffect(() => {
     // Listen for login event from LoginComponent
     const handler = () => setLoggedIn(isLoggedIn());
-    window.addEventListener('login-success', handler);
-    return () => window.removeEventListener('login-success', handler);
+    window.addEventListener("login-success", handler);
+    return () => window.removeEventListener("login-success", handler);
   }, []);
 
   if (!loggedIn) {
@@ -470,20 +508,20 @@ const Root: React.FC = () => {
 // (You can move this logic inside LoginComponent if you prefer)
 const origLoginComponent = LoginComponent;
 (LoginComponent as any) = (props: any) => {
-  const [_, forceUpdate] = React.useReducer(x => x + 1, 0);
+  const [_, forceUpdate] = React.useReducer((x) => x + 1, 0);
   return React.createElement(origLoginComponent, {
     ...props,
     onLoginSuccess: () => {
-      localStorage.setItem('isLoggedIn', 'true');
-      window.dispatchEvent(new Event('login-success'));
+      localStorage.setItem("isLoggedIn", "true");
+      window.dispatchEvent(new Event("login-success"));
       forceUpdate();
-    }
+    },
   });
 };
 
-ReactDOM.render(
+const root = ReactDOM.createRoot(document.getElementById("root")!);
+root.render(
   <StrictMode>
     <Root />
-  </StrictMode>,
-  document.getElementById('root')
+  </StrictMode>
 );
