@@ -1,51 +1,52 @@
-import React, { useState } from 'react';
-import { Button, Typography, Switch, Icon } from 'antd';
+import React, { useState } from "react";
+import { Button, Typography, Switch } from "antd";
+import { CheckOutlined, CloseOutlined, UserOutlined } from "@ant-design/icons";
 
 // Styles
 const popoverContainerStyle: React.CSSProperties = {
   minWidth: 280,
   padding: 28,
-  background: '#232634',
+  background: "#232634",
   borderRadius: 14,
-  boxShadow: '0 4px 32px rgba(0,0,0,0.45)',
-  textAlign: 'center',
-  color: '#f5f6fa',
-  border: '1px solid #2e3244',
-  position: 'relative',
+  boxShadow: "0 4px 32px rgba(0,0,0,0.45)",
+  textAlign: "center",
+  color: "#f5f6fa",
+  border: "1px solid #2e3244",
+  position: "relative",
 };
 
 const logoutButtonStyle: React.CSSProperties = {
-  background: '#e74c3c',
-  borderColor: '#e74c3c',
-  color: '#fff',
+  background: "#e74c3c",
+  borderColor: "#e74c3c",
+  color: "#fff",
   fontWeight: 500,
   borderRadius: 8,
 };
 
 const cancelButtonStyle: React.CSSProperties = {
   marginTop: 10,
-  background: 'transparent',
-  border: '1px solid #35394a',
-  color: '#b0b4c1',
+  background: "transparent",
+  border: "1px solid #35394a",
+  color: "#b0b4c1",
   borderRadius: 8,
 };
 
 const dividerStyle: React.CSSProperties = {
-  margin: '20px 0 16px 0',
-  borderTop: '1px solid #35394a',
+  margin: "20px 0 16px 0",
+  borderTop: "1px solid #35394a",
 };
 
 const themeSwitchStyle: React.CSSProperties = {
   marginLeft: 8,
   marginTop: 16,
-  display: 'inline-block',
+  display: "inline-block",
 };
 
 // Add these props to the component's props type/interface:
 interface UserProfileProps {
   onLogout: () => void;
   onThemeToggle?: () => void;
-  theme?: 'dark' | 'light';
+  theme?: "dark" | "light";
 }
 
 interface UserPopoverProps {
@@ -53,17 +54,21 @@ interface UserPopoverProps {
   onLogout: () => void;
   onCancel: () => void;
   onThemeToggle?: () => void;
-  theme?: 'dark' | 'light';
+  theme?: "dark" | "light";
 }
 
 // Update the component signature:
-const UserProfile: React.FC<UserProfileProps> = ({ onLogout, onThemeToggle, theme }) => {
+const UserProfile: React.FC<UserProfileProps> = ({
+  onLogout,
+  onThemeToggle,
+  theme,
+}) => {
   const [visible, setVisible] = useState(false);
 
   // You can fetch/display real user info here if available
   const userInfo = {
-    name: 'User',
-    email: 'user@email.com',
+    name: "User",
+    email: "user@email.com",
   };
 
   const handleOpen = () => setVisible(true);
@@ -77,37 +82,44 @@ const UserProfile: React.FC<UserProfileProps> = ({ onLogout, onThemeToggle, them
     theme,
   }) => (
     <div style={popoverContainerStyle}>
-      <Typography.Text strong style={{ fontSize: 18, color: '#f5f6fa' }}>{userInfo.name}</Typography.Text>
+      <Typography.Text strong style={{ fontSize: 18, color: "#f5f6fa" }}>
+        {userInfo.name}
+      </Typography.Text>
       <br />
-      <Typography.Text type="secondary" style={{ fontSize: 14, color: '#b0b4c1' }}>{userInfo.email}</Typography.Text>
+      <Typography.Text
+        type="secondary"
+        style={{ fontSize: 14, color: "#b0b4c1" }}
+      >
+        {userInfo.email}
+      </Typography.Text>
       <div style={dividerStyle} />
-      <div style={{ marginBottom: 16, fontSize: 15, color: '#b0b4c1' }}>Do you want to log out?</div>
+      <div style={{ marginBottom: 16, fontSize: 15, color: "#b0b4c1" }}>
+        Do you want to log out?
+      </div>
       <Button
         type="primary"
         block
         style={logoutButtonStyle}
-        onClick={() => { onCancel(); onLogout(); }}
+        onClick={() => {
+          onCancel();
+          onLogout();
+        }}
       >
         Log out
       </Button>
-      <Button
-        block
-        style={cancelButtonStyle}
-        onClick={onCancel}
-      >
+      <Button block style={cancelButtonStyle} onClick={onCancel}>
         Cancel
       </Button>
       {onThemeToggle && (
         <div style={themeSwitchStyle}>
           <Switch
-            checkedChildren={<Icon type="check" />}
-            unCheckedChildren={<Icon type="close" />}
-            checked={theme === 'dark'}
+            checkedChildren={<CheckOutlined />}
+            unCheckedChildren={<CloseOutlined />}
+            checked={theme === "dark"}
             onChange={onThemeToggle}
-            defaultChecked
           />
-          <span style={{ marginLeft: 8, color: '#b0b4c1', fontSize: 14 }}>
-            {theme === 'dark' ? 'Light' : 'Dark'} Theme
+          <span style={{ marginLeft: 8, color: "#b0b4c1", fontSize: 14 }}>
+            {theme === "dark" ? "Light" : "Dark"} Theme
           </span>
         </div>
       )}
@@ -115,40 +127,47 @@ const UserProfile: React.FC<UserProfileProps> = ({ onLogout, onThemeToggle, them
   );
 
   return (
-    <div className="top-bar" style={{ display: 'flex', justifyContent: 'space-between', alignItems: 'center' }}>
+    <div
+      className="top-bar"
+      style={{
+        display: "flex",
+        justifyContent: "space-between",
+        alignItems: "center",
+      }}
+    >
       {/* ...existing code for left/middle of top bar... */}
-      <div style={{ marginLeft: 'auto' }}>
+      <div style={{ marginLeft: "auto" }}>
         <Button
           shape="circle"
-          icon="user"
+          icon={<UserOutlined />}
           onClick={handleOpen}
           style={{
-            background: '#232634',
-            border: '1px solid #35394a',
-            color: '#b0b4c1',
+            background: "#232634",
+            border: "1px solid #35394a",
+            color: "#b0b4c1",
           }}
         />
       </div>
       {visible && (
         <div
           style={{
-            position: 'fixed',
+            position: "fixed",
             zIndex: 9999,
             top: 0,
             left: 0,
-            width: '100vw',
-            height: '100vh',
-            background: 'rgba(23, 25, 34, 0.85)',
-            display: 'flex',
-            alignItems: 'center',
-            justifyContent: 'center',
-            transition: 'background 0.2s',
+            width: "100vw",
+            height: "100vh",
+            background: "rgba(23, 25, 34, 0.85)",
+            display: "flex",
+            alignItems: "center",
+            justifyContent: "center",
+            transition: "background 0.2s",
           }}
           onClick={handleClose}
         >
           <div
-            style={{ pointerEvents: 'auto' }}
-            onClick={e => e.stopPropagation()}
+            style={{ pointerEvents: "auto" }}
+            onClick={(e) => e.stopPropagation()}
           >
             <UserPopover
               userInfo={userInfo}
